requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai==1.40.3
pymongo==4.8.0